+++ packages/sdks/snippets/angular-16-ssr/src/app/app.module.ts
@@ -20,7 +20,6 @@ import { customChildResolver } from './custom-child/custom-child.resolver';
 import { editableRegionsResolver } from './editable-regions/editable-regions-resolver';
 import { EditableRegionComponent } from './editable-regions/editable-regions.component';
 import { LivePreviewComponent } from './live-preview/live-preview.component';
-import { livePreviewResolver } from './live-preview/live-preview.resolver';
 import { NavBarComponent } from './nav-bar/nav-bar.component';
 import { navBarResolver } from './nav-bar/nav-bar.resolver';
 import { ProductDetailsComponent } from './product-details/product-details.component';
@@ -79,7 +78,6 @@ import { productEditorialResolver } from './product-editorial/product-editorial.
       {
         path: 'live-preview',
         component: LivePreviewComponent,
-        resolve: { content: livePreviewResolver },
       },
       {
         path: '**',