+++ packages/sdks/snippets/gen1-react/src/routes/query-cheatsheet/queries/elemMatch.tsx
@@ -1,0 +1,8 @@
+import { builder } from '@builder.io/react';
+import type { GetContentOptions } from '@builder.io/sdk';
+
+export const getProduct = (options: GetContentOptions) =>
+  builder.get('contact-record', {
+    query: { 'data.regions': { $elemMatch: { location: 'US West' } } },
+    ...options,
+  });