+++ packages/sdks/snippets/gen1-react/src/routes/query-cheatsheet/queries/nor.tsx
@@ -1,0 +1,12 @@
+import { builder } from '@builder.io/react';
+import type { GetContentOptions } from '@builder.io/sdk';
+
+export const getProduct = (options: GetContentOptions) =>
+  builder.get('product', {
+    query: {
+      'data.price': {
+        $nor: [{ $lt: 500 }, { $gt: 1000 }],
+      },
+    },
+    ...options,
+  });