+++ packages/sdks/snippets/react/src/routes/query-cheatsheet/queries/lte.tsx
@@ -1,0 +1,7 @@
+import { fetchOneEntry, GetContentOptions } from '@builder.io/sdk-react';
+
+export const getProduct = (options: GetContentOptions) =>
+  fetchOneEntry({
+    query: { 'data.price': { $lte: 455 } },
+    ...options,
+  });